requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.0
numpy==1.24.3
yfinance==0.2.32
scipy==1.11.3
scikit-learn==1.3.2
groq==0.4.2
python-dotenv==1.0.0
streamlit==1.28.1
requests==2.31.0
plotly==5.17.0
pydantic==2.4.2
pydantic[email]==2.4.2
bcrypt==4.1.0
PyJWT==2.8.1
python-multipart==0.0.6
